feat: fixed some missing attributes.
After discussion with Raymond, I'll probably be holding off on further work on this test suite. I'll write tests for the system as I get closer to chapter 9. The reasoning for this is that as I'm still learning the architecture of the messages that need to be passed, any small tests I'm writing are unit tests and not real integration tests. Tests need to be written which are more applicable to the actual messages being sent by the program.